minor tweaks

+++ FlatButton/UIImage+JSFlatButton.m
@@ -30,10 +30,10 @@
 
 - (UIImage *)imageMaskWithColor:(UIColor *)maskColor shadowOffset:(CGPoint)shadow
 {
-    float shadowOpacity = 0.54;
-    BOOL hasShadow = !(shadow.x == 0 && shadow.y == 0);
+    CGFloat shadowOpacity = 0.54f;
+    BOOL hasShadow = !(shadow.x == 0.0f && shadow.y == 0.0f);
     
-    CGRect imageRect = CGRectMake(0, 0, self.size.width, self.size.height);
+    CGRect imageRect = CGRectMake(0.0f, 0.0f, self.size.width, self.size.height);
     CGRect shadowRect = CGRectMake(shadow.x, shadow.y, self.size.width, self.size.height);
 
     CGRect newRect = hasShadow ? CGRectUnion(imageRect, shadowRect) : imageRect;
@@ -41,8 +41,8 @@
     UIGraphicsBeginImageContextWithOptions(newRect.size, NO, self.scale);
     CGContextRef ctx = UIGraphicsGetCurrentContext();
     
-    CGContextScaleCTM(ctx, 1, -1);
-    CGContextTranslateCTM(ctx, 0, -(newRect.size.height));
+    CGContextScaleCTM(ctx, 1.0f, -1.0f);
+    CGContextTranslateCTM(ctx, 0.0f, -(newRect.size.height));
     CGContextSaveGState(ctx);
     
     if(hasShadow) {
